extract JPEG2000 boxes processing into its own file
