+++ libheif/codecs/decoder.cc
@@ -34,6 +34,7 @@
 #include "avif_boxes.h"
 #include "hevc_boxes.h"
 #include "jpeg_boxes.h"
+#include "jpeg2000_boxes.h"
